Simplified ObjectInventory
