+++ Assets/Scripts/Actor/ActorInventory.cs
@@ -7,6 +7,6 @@ public class ActorInventory : ObjectInventory<ActorData, ActorList>, ICourtRecor
     /// <returns>The actor data as an ICourtRecordObject</returns>
     public ICourtRecordObject GetObjectAtIndex(int index)
     {
-        return ObjectStorage[index];
+        return this[index];
     }
 }
