--- Assets/Scripts/ObjectDictionary/OrderedObjectStorage.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 451f8e47526c43249f634caf3bddcba2
-timeCreated: 1631701157
