--- Assets/Tests/EditModeTests/OrderedObjectStorageTests.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1bf0fb62abaa4ebb9d31be5a93715c39
-timeCreated: 1631136484
